feat: add `eject`, which allows to extract the validators from `scalus-examples` into separate sbt projects for easy showcase.
