+++ scalus-examples/jvm/src/main/scala/scalus/examples/factory/Factory.scala
@@ -1,4 +1,4 @@
-package scalus.patterns
+package scalus.examples.factory
 
 import scalus.*
 import scalus.uplc.builtin.Builtins
